Split up files, removed unused code
